+++ CHANGELOG.md
@@ -23,6 +23,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `SendAudioRequest`
   - `SendDocumentRequest`
   - `SendVoiceRequest`
+- Property `ICaptionInlineQueryResult.ParseMode`
+- Property `ParseMode` to inline query results with a caption
+  - `InlineQueryResultPhoto`
+  - `InlineQueryResultGif`
+  - `InlineQueryResultCachedMpeg4Gif`
+  - `InlineQueryResultVideo`
+  - `InlineQueryResultAudio`
+  - `InlineQueryResultVoice`
+  - `InlineQueryResultDocument`
+  - `InlineQueryResultCachedPhoto`
+  - `InlineQueryResultCachedGif`
+  - `InlineQueryResultCachedMpeg4Gif`
+  - `InlineQueryResultCachedDocument`
+  - `InlineQueryResultVideo`
+  - `InlineQueryResultCachedVoice`
+  - `InlineQueryResultCachedAudio`
 - Property `InputMediaBase.ParseMode`
 - Property `SendVideoRequest.SupportsStreaming`
 - Property `InputMediaVideo.SupportsStreaming`